tasklist test

+++ app/Http/Controllers/TasksController.php
@@ -15,12 +15,55 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-        
+/**
+        // メッセージ一覧を取得
+        $tasks = task::all();
+
+        // メッセージ一覧ビューでそれを表示
         return view('tasks.index', [
             'tasks' => $tasks,
-            ]);
+        ]);
     }
+ */     
+      
+       
+       $data = [];
+       
+       /**
+        * (\Auth::check()は、ユーザがログインしているか
+        * どうかを調べるための関数
+        */
+        if (\Auth::check()) { // 認証済みの場合
+        
+        
+        // \Auth::user();ログイン中のユーザを取得
+        $user = \Auth::user();
+       
+       /**
+        * orderBy(created_at,'desc')
+        * created_atのカラムを基準にdesc降順で表示
+        * 
+       */
+        $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+       
+        //dd($tasks);
+        
+       
+        $data = [
+              'user' => $user,
+              'tasks' => $tasks,
+             ];
+         
+        //  resources\views\tasks\index
+         
+        return view('tasks.index', $data,);
+    
+      }else {
+         //ログインしていない場合、ログイン画面にリダイレクト
+          return redirect('login');
+      }
+    } 
+
 
     /**
      * Show the form for creating a new resource.
@@ -47,15 +90,20 @@ class TasksController extends Controller
        // バリデーション
         $request->validate([
             
-            'status' => 'required|max:10',
-            'content' => 'required'
+            
+            'content' => 'required',
+            'status' => 'required|max:10'
         ]);
         
-      // メッセージを作成
-        $task = new Task;
-        $task->status = $request->status;    // 追加
-        $task->content = $request->content;
-        $task->save();
+        
+         // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+          'content' => $request->content,
+          'status' => $request->status,
+        ]);
+
+        
+     
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -107,7 +155,7 @@ class TasksController extends Controller
         $request->validate([
             
              'status' => 'required|max:10',
-            'content' => 'required|',
+             'content' => 'required|',
         ]);
         
         // idの値でメッセージを検索して取得
@@ -131,11 +179,15 @@ class TasksController extends Controller
     public function destroy($id)
     {
           // idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
-
+        $task = \App\Task::findOrFail($id);
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+           $task->delete();
+       } 
         // トップページへリダイレクトさせる
-        return redirect('/');
+    
+       return redirect('/');
+    
     }
 }
